qti-generator/ai-text/requirements.txt
# Claude API
anthropic>=0.40.0

# RDF/LOD処理
rdflib>=7.0.0

# HTML解析（commentary取得用）
beautifulsoup4>=4.12.0
requests>=2.31.0

# XML処理
lxml>=5.0.0

# ユーティリティ
python-dotenv>=1.0.0
pydantic>=2.0.0

# 開発・テスト
pytest>=8.0.0
pytest-asyncio>=0.23.0
